Refactoring of existing methods and tests
