+++ lib/game_mechanics.rb
@@ -1,4 +1,5 @@
 module GameMechanics
+  
   def rand_number
     rand(0..100)
   end
