+++ lib/ruby_server.rb
@@ -7,10 +7,10 @@ class RubyServer
   attr_reader :server,
               :port,
               :hello_requests,
-              :all_requests
+              :all_requests,
+              :game
 
-  include ParseAndFormat
-  include Response
+  include ParseAndFormat, Response
 
   def initialize(port = 9292)
     @server = TCPServer.new(port)
@@ -18,6 +18,14 @@ class RubyServer
     @hello_requests = -1
     @all_requests = 0
   end
+
+  def keep_open
+    while true
+      process_request
+    end
+  end
+
+  private
 
   def process_request
     client = server.accept
@@ -44,79 +52,13 @@ class RubyServer
     received
   end
 
-  def headers(output, responde_code = '200 ok', location = "")
-    headers = ["http/1.1 #{responde_code}",
-              "location: #{location}",
-              "date: #{Time.now.strftime('%a, %e %b %Y %H:%M:%S %z')}",
-              'server: ruby',
-              'content-type: text/html; charset=iso-8859-1',
-              "content-length: #{output.length}\r\n\r\n"].join("\r\n")
-  end
-
-  def keep_open
-    while true
-      process_request
-    end
-  end
-
-  # def set_response_from_path(lines)
-  #   if requested_path(lines) == '/'
-  #     "<pre>" + debugger(lines) + "</pre>"
-  #   elsif requested_path(lines) == '/hello'
-  #     @hello_requests += 1
-  #     "Hello, World! (#{hello_requests})"
-  #   elsif requested_path(lines) == '/datetime'
-  #     Time.now.strftime("%I:%M%p on %A, %B %d, %Y")
-  #   elsif requested_path(lines)[0..11] == '/word_search'
-  #     "#{find_word(lines).upcase} is #{word_search(find_word(lines))} word"
-  #   elsif requested_path(lines) == '/shutdown'
-  #     server.close
-  #     "Total Requests: #{all_requests}"
-  #   else
-  #     set_response_for_game_paths(lines)
-  #   end
-  # end
-  #
-  # def set_response_for_game_paths(lines)
-  #   if requested_path(lines) == '/start_game' && post?(lines)
-  #     create_game
-  #   elsif requested_path(lines) == '/game' && get?(lines)
-  #     if active_game?
-  #       @game.status + "\n\n" +
-  #       @game.compare_guess(@game.secret_number, @game.guess)
-  #     else
-  #       "You haven't started a game yet. POST to /start_game first."
-  #     end
-  #   elsif requested_path(lines)[0..4] == '/game' && post?(lines)
-  #     if active_game?
-  #       @game.num_guesses += 1
-  #     else
-  #       "You haven't started a game yet. POST to /start_game first."
-  #     end
-  #   else
-  #     "404 Not Found"
-  #   end
-  # end
-
-  # def active_game?
-  #   defined?(@game) != nil
-  # end
-
-  def check_and_set_guess(lines, user_guess)
-    @game.guess = user_guess if active_game? && post?(lines)
-  end
-
   def create_game
     @game = Game.new
-    @game.start_game
+    game.start_game
   end
 
-  def post?(lines)
-    check_type_of_request(lines) == "POST"
-  end
-
-  def get?(lines)
-    check_type_of_request(lines) == "GET"
+  def check_and_set_guess(lines, user_guess)
+    game.guess = user_guess if active_game? && post?(lines)
   end
 
 end
